move tests to JavaBinaryInteropTest
